requirements.txt
selenium==4.15.2
webdriver_manager==4.0.1
pyyaml==6.0.1
urllib3==2.0.7
